+++ game_patch/multi/multi.cpp
@@ -15,6 +15,7 @@
 #include "multi_private.h"
 #include "alpine_packets.h"
 #include "sprays.h"
+#include "kill_attribution.h"
 #include "server_internal.h"
 #include "gametype.h"
 #include "rounds.h"
@@ -1326,6 +1327,7 @@ void multi_do_patch()
     multi_customize_listen_server_settings_patch.install();
 
     multi_kill_do_patch();
+    kill_attribution_do_patch();
     sprays_do_patch();
     faction_files_do_patch();
     level_download_do_patch();